Fixed errors
I think we need to refactor the intake subsystem. There are just too many issues with our current implementation
